Parse and serve image dimensions/color/etc from S3 object metadata
Making an individual HEAD request for each image is excruciatingly slow,
but as a quick-and-dumb implementation it's fine for now - we cache
metadata indefinitely so we only pay the cost on startup and when new
images are uploaded, and this is a simple approach that keeps image
metadata stored with the images themselves. We can optimize later (so
startup time won't scale linearly with the number of images).